refactor: rename `erlab.analysis.utilities` to `erlab.analysis.utils`
